+++ src/skeleton.c
@@ -5,11 +5,11 @@
 #include <R_ext/Rdynload.h>
 
 void F77_NAME(s_brook90_f)(double *siteparam, double *climveg, double *param, double *pdur,
-    double *soil_materials, double *soil_nodes, double *precdat,
+    double *soil_materials, double *soil_nodes, double *precdat, double *gwt_depth,
     int *pr, int *timer, int *error, int *chk, double *output, double *output_layer);
 
 extern SEXP s_brook90_c(SEXP siteparam, SEXP climveg, SEXP param, SEXP pdur, SEXP soil_materials,
-    SEXP soil_nodes, SEXP precdat, SEXP pr, SEXP timer, SEXP chk, SEXP n_days, SEXP n_lays, SEXP n_pint){
+    SEXP soil_nodes, SEXP precdat,SEXP gwt_depth, SEXP pr, SEXP timer, SEXP chk, SEXP n_days, SEXP n_lays, SEXP n_pint){
 
 
     int n; // total length of layer output
@@ -34,8 +34,8 @@ extern SEXP s_brook90_c(SEXP siteparam, SEXP climveg, SEXP param, SEXP pdur, SEX
     setAttrib( output_layer, R_DimSymbol, dims);
 
     F77_CALL(s_brook90_f)(REAL(siteparam), REAL(climveg), REAL(param), REAL(pdur),
-             REAL(soil_materials),REAL(soil_nodes), REAL(precdat), INTEGER(pr),
-             INTEGER(timer), INTEGER(error),INTEGER(chk),
+             REAL(soil_materials),REAL(soil_nodes), REAL(precdat),REAL(gwt_depth),
+             INTEGER(pr),INTEGER(timer), INTEGER(error),INTEGER(chk),
              REAL(output), REAL(output_layer));
 
     SEXP out_full = PROTECT( allocVector(VECSXP, 3) );
@@ -49,7 +49,7 @@ extern SEXP s_brook90_c(SEXP siteparam, SEXP climveg, SEXP param, SEXP pdur, SEX
 }
 
 static const R_CallMethodDef CallEntries[] = {
-  {"s_brook90_c",   (DL_FUNC) &s_brook90_c,   13},
+  {"s_brook90_c",   (DL_FUNC) &s_brook90_c,   14},
   {NULL,         NULL,                0}
 };