+++ NEWS.md
@@ -2,7 +2,8 @@
 
 ## Changes
  - removed default value for `topsoil` arguments in functions `hydpar_hypres()` and `hydpar_hypres_tab()`.
-
+ - simplified the low-level interface function, bringing a breaking change to users using the `r_lwfbrook90()` instead of `run_LWFB90()` for running the model
+ - added an option to use a timeseries of groundwater table depths in the simulation. For activating this option, just provide a data.frame with columns 'dates', 'water_table_depth' to the `param_b90$water_table_depth` list item instead of a single fixed value.
 
 # LWFBrook90R 0.6.2